refactor(cubesql): Split QueryPlanner and QueryPlan  (#8584)
