+++ rust/cubesql/cubesql/src/compile/rewrite/rules/wrapper/subquery.rs
@@ -1,13 +1,11 @@
 use crate::{
-    compile::{
-        rewrite::{
-            analysis::LogicalPlanAnalysis, cube_scan_wrapper, empty_relation,
-            rules::wrapper::WrapperRules, transforming_rewrite, wrapper_pullup_replacer,
-            wrapper_pushdown_replacer, EmptyRelationDerivedSourceTableName, LogicalPlanLanguage,
-            WrapperPullupReplacerAliasToCube,
-        },
-        MetaContext,
+    compile::rewrite::{
+        analysis::LogicalPlanAnalysis, cube_scan_wrapper, empty_relation,
+        rules::wrapper::WrapperRules, transforming_rewrite, wrapper_pullup_replacer,
+        wrapper_pushdown_replacer, EmptyRelationDerivedSourceTableName, LogicalPlanLanguage,
+        WrapperPullupReplacerAliasToCube,
     },
+    transport::MetaContext,
     var, var_iter, var_list_iter,
 };
 use egg::{EGraph, Rewrite, Subst, Var};
